Pages + Nav bard + linked style sheet
So I’ve made each of the pages that we had yet to make - they each atm
just contain a nav bar, link to the style sheet and the title; “Eggs”
On the Omelette page to carousel seems to sit over the nav bar
preventing you from being able to click on it so I’ll see if this is
something I can move down.
I think Sarah - you’d put the carousel of omelettes in the scrambled
page so I’ve renamed it.
I also move the style sheet to the same folder as the pages as I was
having trouble linking it.
x